backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.1
PyJWT==2.9.0
email-validator==2.2.0
